+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 **Helm**
 - Add `atomic`, `version`, and `createNS` parameters to appropriate methods.
 - Add `registry_login` method.
+- Convert `version` parameter to optional in `rollback` method.
 
 **Packer**
 - Revert code to Groovy 2.4 compatible.